% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wk_07.R
\name{AirPollution}
\alias{AirPollution}
\title{SO2 air pollution by EU city.}
\usage{
AirPollution(COUNTRY, LEVEL)
}
\arguments{
\item{COUNTRY}{Country of interest, use code (2 letters) chosen from EUSO2 dataset}

\item{LEVEL}{Concentration of SO2 pollution (ug/m^3) as threshold level}
}
\value{
OUTPUT Number of cities within the country exceeding threshold level
}
\description{
This function summarizes the number of of cities that have air pollution levels
 in micrograms/cubic meter above a certain level.
}
\examples{
COUNTRY = DE
LEVEL = 20
AirPollution(COUNTRY, LEVEL)
}
\author{
Mitchell, Steph, Elise
}

