% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumCities.R
\name{NumCities}
\alias{NumCities}
\title{SO2 air pollution in EU cities (2013)}
\usage{
NumCities(COUNTRY, LEVEL, POLLUTION)
}
\arguments{
\item{COUNTRY}{Country of interest, use code (2 letters) chosen from EUSO2 dataset}

\item{LEVEL}{Concentration of SO2 pollution (ug/m^3) as threshold level}

\item{POLLUTION}{is the pollution type (SO2,PM10, PM2.5,NO2,O3)}
}
\value{
OUTPUT Number of cities within the country exceeding threshold level
}
\description{
This function summarizes the number of of cities that have air pollution levels
 in micrograms/cubic meter above a certain level.
}
\examples{
COUNTRY = 'DE'
LEVEL = 10
POLLUTIOn='SO2'
NumCities(COUNTRY, LEVEL,POLLUTION)
}
\author{
Mitchell, Steph, Elise
}

